2D Screen Resolution appears to be scaling correctly (this is always a problem)
Scaling the 2D screen resolution always appears to be a problem in Unity, its something I've yet to figure out how to solve perfectly.

It was the first problem I tried to look at though and what I did was first set the Player Settings to only allow Portrait. Then in the UI elements I set the Canvas Scaler to “Match with Height” and the "Reference Resolution" as 100 : 200, and the "Match" value to 0.5. I then only used UI element components (eg. Image and Button), and it appeared to scale relatively well. I also set the camera's background colour to be the same as my background image, so for now it appears to be scaling relatively well. The real test will come when we actually get a video running on the app.