Added a VideoPlayer component and hooked it up
Simple change here, I just added a Video Player component and hooked up a video from within the project.

I was very happy to find that it re-adjusted very well in all the different resolutions. Next step is to test it out with a url from youtube, I think that's the only unknown in this project, everything else will be polish and building the logic.