Made a basic initial flow for the actual app
Setup a basic flow where you have a Start screen, that you click through to the Main Menu. From there you have the option of Beginner, Intermediate, Advanced videos and when you select one of them you can select the first video (which just leads to the same video as before). This little flow all happens now in the AppDirector though and the scene is a lot cleaner. Still have an odd issue about ordering with the loading screen, as it loads a behind my scene.