Got flow looking like Sean's initial prototype (except for videos)
Made all the buttons of the app like in Sean's design and got the flow of the app as Sean had suggested. Just missing the videos that he thought would be ideal when opening the app and selecting a category. We also have the first video that we can work with, which I need to save to a private youtube channel.