Transferred from YoutubePlayer to VimeoPlayer
I've moved to using Vimeo as it appeared to give better streaming performance than Youtube. VimeoPlayer now does the heavy lifting of downloading and streaming, and although its not coded in a very user-friendly way, it seems to give a better result. My main concern is still about people with low internet bandwidth as the video appears choppy if you have poor internet.